refactor: 패키지 이동 및 사용하지 않는 XML, Fragment 제거
