+++ android/app/src/main/java/com/yagubogu/presentation/login/LoginActivity.kt
@@ -36,7 +36,6 @@ import com.yagubogu.presentation.login.auth.GoogleCredentialManager
 import com.yagubogu.presentation.login.model.InAppUpdateType
 import com.yagubogu.presentation.login.model.LoginResult
 import com.yagubogu.presentation.login.model.VersionInfo
-import com.yagubogu.presentation.setting.SettingActivity
 import com.yagubogu.presentation.util.showSnackbar
 import com.yagubogu.presentation.util.showToast
 import com.yagubogu.ui.main.MainActivity
