+++ android/app/src/main/java/com/yagubogu/ui/setting/component/dialog/SettingDialog.kt
@@ -5,7 +5,7 @@ import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import com.yagubogu.presentation.setting.SettingViewModel
+import com.yagubogu.ui.setting.SettingViewModel
 import com.yagubogu.ui.setting.component.model.SettingDialogEvent
 
 @Composable
