+++ android/app/src/main/java/com/yagubogu/ui/setting/component/model/PresignedUrlCompleteItem.kt
@@ -1,4 +1,4 @@
-package com.yagubogu.presentation.setting
+package com.yagubogu.ui.setting.component.model
 
 data class PresignedUrlCompleteItem(
     val imageUrl: String,
