+++ android/app/src/main/java/com/yagubogu/ui/setting/component/model/PresignedUrlItem.kt
@@ -1,4 +1,4 @@
-package com.yagubogu.presentation.setting
+package com.yagubogu.ui.setting.component.model
 
 data class PresignedUrlItem(
     val key: String,
