--- android/app/src/main/res/layout/activity_setting.xml
@@ -1,52 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools">
-
-    <data>
-
-        <variable
-            name="viewModel"
-            type="com.yagubogu.presentation.setting.SettingViewModel" />
-    </data>
-
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:id="@+id/constraint_activity_setting_root"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:background="@color/gray050"
-        tools:context=".presentation.setting.SettingActivity">
-
-        <ImageView
-            android:id="@+id/iv_arrow_left"
-            android:layout_width="24dp"
-            android:layout_height="24dp"
-            android:layout_marginStart="20dp"
-            android:background="?attr/selectableItemBackgroundBorderless"
-            android:src="@drawable/ic_arrow_left"
-            app:layout_constraintBottom_toBottomOf="@id/tv_setting"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="@id/tv_setting" />
-
-        <TextView
-            android:id="@+id/tv_setting"
-            style="@style/TextView.Pretendard.Bold.20"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginVertical="12dp"
-            android:text="@{viewModel.settingTitle}"
-            app:layout_constraintBottom_toTopOf="@id/fcv_setting"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent"
-            tools:text="설정" />
-
-        <androidx.fragment.app.FragmentContainerView
-            android:id="@+id/fcv_setting"
-            android:layout_width="match_parent"
-            android:layout_height="0dp"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/tv_setting" />
-
-    </androidx.constraintlayout.widget.ConstraintLayout>
-</layout>
