--- android/app/src/main/res/layout/fragment_setting_account.xml
@@ -1,40 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools">
-
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        tools:context=".presentation.setting.SettingAccountFragment">
-
-        <androidx.constraintlayout.widget.ConstraintLayout
-            android:id="@+id/constraint_setting_account"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginHorizontal="20dp"
-            android:layout_marginTop="8dp"
-            android:background="@drawable/bg_white_radius_12dp"
-            android:padding="20dp"
-            app:layout_constraintTop_toTopOf="parent">
-
-            <include
-                android:id="@+id/layout_logout"
-                layout="@layout/item_setting_menu"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                app:layout_constraintTop_toTopOf="parent"
-                app:menuTitle="@{@string/setting_logout}" />
-
-            <include
-                android:id="@+id/layout_delete_account"
-                layout="@layout/item_setting_menu"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:layout_marginTop="20dp"
-                app:layout_constraintTop_toBottomOf="@id/layout_logout"
-                app:menuTitle="@{@string/setting_delete_account}" />
-
-        </androidx.constraintlayout.widget.ConstraintLayout>
-    </androidx.constraintlayout.widget.ConstraintLayout>
-</layout>
